Temporary fix for small devices
