+++ app/Response/Components/AsssignmentDescriptions/index.ts
@@ -1,0 +1,4 @@
+export * from "./PostDescription";
+export * from "./VideoDescription";
+export * from "./WebinarSpeakerDescription";
+export * from "./InvitesDescription";
