+++ app/Response/Components/InterestDescriptionText.tsx
@@ -1,0 +1,14 @@
+import { Typography } from "@mui/material";
+
+//Mark: InterestDescriptionText
+export function InterestDescriptionText() {
+    return (
+        <Typography>
+            Wir haben Sie aufgrund ihrer Expertise und ihres Interesses an den Themen des Kunden
+            ausgewählt. Bitte teilen Sie uns mit, ob Sie an der Kampagne teilnehmen möchten.
+            <br />
+            Wir werden Ihre Entscheidung an unseren Kunden weiterleiten und Sie zeitnah über deren
+            Auswahl informieren.
+        </Typography>
+    );
+}
