+++ app/Response/Components/Title.tsx
@@ -1,0 +1,11 @@
+import { Box, Typography } from "@mui/material";
+import { SwinxLogo } from "./SwinxLogo";
+
+export default function Title() {
+    return (
+        <Box id="Title">
+            <SwinxLogo />
+            <Typography variant="h4">Werden Sie Teil unserer Kampagne</Typography>
+        </Box>
+    );
+}
