+++ app/Response/Functions/displayCountryString.tsx
@@ -1,0 +1,5 @@
+import { Nullable } from "@/app/Definitions/types";
+
+export function displayCountryString(countries: Nullable<string>[]) {
+    return countries.join(", ").replace(/, ([^,]*)$/, " und $1");
+}
